Emulation of 32 bit Windows applications is still unsupported.

+++ src/analyzer/object_watching.hpp
@@ -10,7 +10,7 @@ emulator_hook* watch_object(windows_emulator& emu, const std::set<std::string, s
     const reflect_type_info<T> info{};
 
     return emu.emu().hook_memory_read(
-        object.value(), object.size(),
+        object.value(), static_cast<size_t>(object.size()),
         [i = std::move(info), object, &emu, cache_logging, modules](const uint64_t address, const void*, size_t) {
             const auto rip = emu.emu().read_instruction_pointer();
             const auto* mod = emu.mod_manager.find_by_address(rip);
@@ -33,6 +33,7 @@ emulator_hook* watch_object(windows_emulator& emu, const std::set<std::string, s
             const auto offset = address - object.value();
             emu.log.print(is_main_access ? color::green : color::dark_gray,
                           "Object access: %s - 0x%llX (%s) at 0x%llX (%s)\n", i.get_type_name().c_str(), offset,
-                          i.get_member_name(offset).c_str(), rip, mod ? mod->name.c_str() : "<N/A>");
+                          i.get_member_name(static_cast<size_t>(offset)).c_str(), rip,
+                          mod ? mod->name.c_str() : "<N/A>");
         });
 }
